+++ CHANGELOG.md
@@ -65,6 +65,10 @@
         * Whenever you attempt a dispel check based on your caster level, you gain a +2 bonus on the check.
     * Greater Dispel Focus
         * Whenever you attempt a dispel check based on your caster level, you gain a +2 bonus to the check. This stacks with the bonus from Dispel Focus.
+    * Intensified Spell (Metamagic)
+        * An intensified spell increases the maximum number of damage dice by 5 levels. You must actually have sufficient caster levels to surpass the maximum in order to benefit from this feat. No other variables of the spell are affected, and spells that inflict damage that is not modified by caster level are not affected by this feat.
+    * Rime Spell (Metamagic)
+        * The frost of your cold spell clings to the target, impeding it for a short time. A rime spell causes creatures that takes cold damage from the spell to become entangled for a number of rounds equal to the original level of the spell.
     * Quicken Blessing
         * Choose one of your blessings that normally requires a standard action to use. You can expend two of your daily uses of blessings to deliver that blessing (regardless of whether it’s a minor or major effect) as a swift action instead.
     * Two-Weapon Defense